Major updates
This commit is made uo of changes of a whole day. Most of the changes
too place on 'document.py' which incvolved refactoring some code into
new module 'resource.py'. There are lot of changes in this commit which
are intended to improve this project.

Changes in this commit are foundation for avoiding problems experienced
while working with 'navaly'(another project). document.py has recived major
changes which include some of its code moved to new module resource.py.

Some methods from classes defined in document.py were changed to attributes
to help ease access to their data. New methods were also added including being
able to get modification time of file in path.

One important method added is source_resembles() which checks if source is similar
to expected source even if it may not be supported. This may be important for checking
if string resembles file path.

pathmod.py was added for paths related functionality inclung validating paths, getting files
and other activities relating to paths. It was splitout from filepath.py when demand
for directory paths increased.

Explaing all changes made to the files is impossible at moment as I am very tired. Last commit
was never commented due to same reason but just commited to call it a day for that day.
Creating long commit messages takes time which can be used for adding new feature to project.

One thing I dont like doing is using classes, they can get complex very fast if not planned.
Its easy to lock fuction in class while it should also be used outside module or class that
does not inherut the class. My use of classes is poor and most of the time it causes problems
due to too much dependence of the classes.

This project is mostly functional like with classes used in high level in modules like resource.py
and document.py. Almost each class is backed by module that provided implementations for its
methods. The module backing class can be used anywere while method of object is not easy to use
outside without reuiting object or sometimes impossible if it does not accept arguments.

This project will just challenge as navaly did but if it succeed it will open oportunities.