Update: source files
New files
---------
highlevel.py
master.py

This commit ha sbeen concerned with making sure things work well at high level
and developing high level code over resource classes. There has been modifications
to simplify things to ensure that high level code will be able to perform well.

Realise that high level code for this library is never planned and everything added
is just created on fly while writting code. The high level code is more inspired by
that from 'navaly' prject but avoid mistakes it brought to 'navaly'.

Seperate module 'master.py' was added from managing resources classes as similar to
'master_fetch' and 'master_parse' to navaly. 'master.py' contains Master class which
accepts source and then create resource instances from the source.

Thankfully unlike navaly, resource classes and Master do not raise any error when source
is not supported but rather return exception for that unsupported part. Resource class does
not raise exception no matter what resource object is passed to it.

Thats because the initialiser does not call any of instance method. This allows the object to
still be created and be able to use the parts that may still work even when source is
not valid. 'Navaly' would raise exception when creating instance which prevent creating object
in first place or even makes object creating much slow due to method calls.

master.py just helps in creating and maching resource instances. It makes it possible to get supported
resource objects orones resembling source. highlevel.py module was added to be built on top of master
and create high level functions like guess_extension() which may be usefull as api for the library.

highlevel.py behaves as expected and much better than high level code for navaly. It was all thanks to
'resembles' feature added and/or improved in previous commit. It makes it posible to know what source may
fall into if it was valid. File path that does not yet exists does not mean is not file path but is
a file path that needs few lines to guesss if it is really resembling files path.

Instead of raising exceptions, None is retured which means nothing as found or source is not supported.
This prevent the need to define lot of exceptions including even error messages which can take long
to write. Returning None makes them even more usable just like 'mimetypes' which its functions returns
None when content type cannot be found.

Its easy to handle None with if statement but requires more lines to handle exception.