Update: README.md
Some of examples in README.md are not working as expected due to
bugs in source code of project. Since this project is missing tests,
it becomes harder to detect bugs or assure that there is no bugs.

Few bugs have been detected while developing examples which were fixed
in previous commit. SOme bugs are still existing and will be exposed when
corresponding functions are executed.