Bug fix & refactor
Urls on with 'localhost' hosname were considered being on remote even when
they are locally. This was bacause URL class did not implement available_locally()
function.

Code for implementing available_locally() was in 'weburl' module. The code was moved
to 'urlmod' module so that it can be used for any kind of url including file paths
and non web urls.

Finally, available_locally() was implemented.

Issues
------
- URL is taking priority over file paths and directory paths when they are not supported.
  This was expected but it may cause confution when file path does not exists which may
  cause it to be interpreted as url. Its not that bad as file path or directory paths are
  are valid urls.